+++ components/AuthForm/index.js
@@ -1,13 +1,15 @@
-import { useState, useRef } from "react";
-import { signIn } from "next-auth/react";
-import { useRouter } from "next/router";
-
+// Styles Imports
 import { StyledCard } from "../StyledComponents/StyledCard";
 import { StyledForm, StyledFormInput, StyledFormLabel } from "../AddEntryForm";
 import { StyledSettingsButton } from "../StyledComponents/StyledButtons";
 import styled from "styled-components";
 import ErrorIcon from "@/assets/icons/ErrorIcon";
 import { StyledSubmitButtonRight, StyledWarningText } from "../EditingForm";
+
+// Functions and Dependencies Imports
+import { useState, useRef } from "react";
+import { signIn } from "next-auth/react";
+import { useRouter } from "next/router";
 
 function AuthForm() {
   const [registered, setRegistered] = useState(false);
@@ -92,9 +94,7 @@ function AuthForm() {
     <StyledCard>
       {!registered ? (
         <>
-          <h2 className="inherit-background-color">
-            {loginProcess ? "Login" : "Sign Up"}
-          </h2>
+          <h2>{loginProcess ? "Login" : "Sign Up"}</h2>
           <StyledForm onSubmit={submitHandler}>
             <StyledFormLabel htmlFor="username">Username</StyledFormLabel>
             <StyledFormInput
@@ -106,11 +106,7 @@ function AuthForm() {
 
             {loginError.hasError && (
               <StyledErrorMessage>
-                <span
-                  className="inherit-background-color"
-                  role="img"
-                  aria-label="error"
-                >
+                <span role="img" aria-label="error">
                   <ErrorIcon height="18px" width="18px" />
                 </span>
                 {loginError.errorMessage}
@@ -142,9 +138,7 @@ function AuthForm() {
         </>
       ) : (
         <>
-          <p className="inherit-background-color">
-            You have successfully registered!
-          </p>
+          <p>You have successfully registered!</p>
 
           <StyledSettingsButton
             onClick={() => router.reload()}
@@ -158,6 +152,8 @@ function AuthForm() {
   );
 }
 
+// Styles
+
 const StyledLoginText = styled(StyledWarningText)`
   margin-bottom: 3rem;
 `;
@@ -166,10 +162,6 @@ const StyledErrorMessage = styled.p`
   border: 1px solid var(--highlight-red);
   padding: 5px;
   border-radius: 5px;
-
-  &::before {
-    background-color: inherit;
-  }
 `;
 
 const StyledSettingsButtonLeft = styled(StyledSettingsButton)`
