+++ components/BarDiagram/index.js
@@ -1,3 +1,4 @@
+// Styles Imports
 import { styled } from "styled-components";
 
 export default function BarDiagram({ inputArray }) {
@@ -22,6 +23,8 @@ export default function BarDiagram({ inputArray }) {
     );
   }
 }
+
+// Styles
 
 const DiagramLabel = styled.span`
   width: 2.5rem;
