+++ components/SearchComponent/index.js
@@ -1,7 +1,13 @@
+// Styles Imports
 import styled from "styled-components";
-import SearchResults from "../SearchResults";
-import { useEffect } from "react";
 import SearchIcon from "@/assets/icons/SearchIcon";
+
+//Components Imports
+import SearchResults from "../SearchResults";
+
+// Functions and Dependencies Imports
+import { useEffect } from "react";
+import { device } from "@/utils/globalValues";
 
 function SearchComponent({
   query,
@@ -36,47 +42,42 @@ function SearchComponent({
     // Search Dictionary if no searchResults in list
     if (searchResults.length === 0 && query.length > 0) {
       handleSetDictionaryQuery(searchQuery.searchInput.toLowerCase());
-
       handleHasEntries(false);
     }
 
     form.reset();
   }
 
-  // OnChange should still look up the words
+  // OnChange look up the words
   function handleSearchBarOnInput(query) {
-    // Empty useStates before searching for new word
     handleSetDictionaryQuery("");
     handleSearchResults([]);
     handleSetDictionaryResults([]);
 
-    // search in word List
     handleSetQuery(query);
     handleSearchInput(query);
   }
 
   return (
     <>
-      <StyledSearchBarForm onSubmit={(event) => handleSearchBarSubmit(event)}>
-        <StyledSearchBar
-          onChange={(event) =>
-            handleSearchBarOnInput(event.target.value.toLowerCase())
-          }
-          type="text"
-          placeholder='Dictionary Search on "enter"'
-          aria-label="search-bar"
-          name="searchInput"
-        />
-        <StyledSearchBarButton>
-          <span
-            className="inherit-background-color"
-            role="img"
-            aria-label="Search"
-          >
-            <SearchIcon height="30px" width="30px" />
-          </span>
-        </StyledSearchBarButton>
-      </StyledSearchBarForm>
+      <StyledStickyDivLargeMargin>
+        <StyledSearchBarForm onSubmit={(event) => handleSearchBarSubmit(event)}>
+          <StyledSearchBar
+            onChange={(event) =>
+              handleSearchBarOnInput(event.target.value.toLowerCase())
+            }
+            type="text"
+            placeholder='Dictionary Search on "enter"'
+            aria-label="search-bar"
+            name="searchInput"
+          />
+          <StyledSearchBarButton>
+            <span role="img" aria-label="Search">
+              <SearchIcon height="25px" width="25px" />
+            </span>
+          </StyledSearchBarButton>
+        </StyledSearchBarForm>
+      </StyledStickyDivLargeMargin>
 
       <SearchResults
         searchResults={searchResults}
@@ -92,21 +93,29 @@ function SearchComponent({
     </>
   );
 }
-const StyledSearchBarForm = styled.form`
+
+// Styles
+
+export const StyledSearchBarForm = styled.form`
   width: 80%;
   display: flex;
   align-self: center;
   justify-content: space-between;
   background-color: var(--dark-mode-text-color);
   padding: 10px;
+  border: 2px solid var(--light-grey);
   border-radius: 25px;
   z-index: 3;
   box-shadow: var(--default-box-shadow);
   position: sticky;
   top: 5rem;
+
+  @media ${device.tablet} {
+    width: 60%;
+  }
 `;
 
-const StyledSearchBar = styled.input`
+export const StyledSearchBar = styled.input`
   background-color: inherit;
   padding: 10px;
   border: none;
@@ -117,10 +126,19 @@ const StyledSearchBar = styled.input`
   }
 `;
 
-const StyledSearchBarButton = styled.button`
+export const StyledSearchBarButton = styled.button`
   border: none;
   background-color: transparent;
-  padding: 0 15px;
+  padding: 0 11px;
+`;
+
+const StyledStickyDivLargeMargin = styled.div`
+  display: flex;
+  justify-content: center;
+  z-index: 10;
+  position: sticky;
+  top: 3.5rem;
+  width: 100%;
 `;
 
 export default SearchComponent;
