+++ components/StyledComponents/Links.js
@@ -4,3 +4,7 @@ import styled from "styled-components";
 export const LinkWithoutDecoration = styled(Link)`
   text-decoration: none;
 `;
+
+export const NoStyleLink = styled(Link)`
+  text-decoration: none;
+`;
