+++ components/StyledComponents/Modal.js
@@ -4,7 +4,6 @@ export const Modal = styled.main`
   position: fixed;
   top: 0;
   background-color: var(--overlay);
-  /* backdrop-filter: blur(15px); */
   width: 100vw;
   height: 100vh;
   z-index: 6;
@@ -19,4 +18,16 @@ export const FixedCenteredPosition = styled.div`
   top: 50%;
   right: 50%;
   z-index: 5;
+`;
+
+export const StudyModal = styled.section`
+  position: fixed;
+  left: 0;
+  top: 0;
+  background-image: url("/background-vector.png");
+  background-repeat: repeat;
+
+  width: 100vw;
+  height: 100vh;
+  z-index: 6;
 `;
