+++ components/StyledComponents/ParagraphsAndHeadings.js
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const StyledHeading2 = styled.h2`
+  padding: 1.5rem 0 0.3rem 0;
+  color: var(--dark-main);
+  font-size: 1.2rem;
+  margin: 0px;
+`;
+
+export const StyledCardParagraph = styled.p`
+  margin-top: 0;
+`;
+export const StyledHeading3 = styled.h3`
+  display: inline-block;
+  padding: 1.5rem 0 0.3rem 0;
+  color: var(--dark-main);
+  font-size: 1.2rem;
+  margin: 0px;
+`;
