+++ fonts.js
@@ -1,9 +1,4 @@
-import { Permanent_Marker, Poppins } from "next/font/google";
-
-export const permanent = Permanent_Marker({
-  subsets: ["latin"],
-  weight: ["400"],
-});
+import { Poppins } from "next/font/google";
 
 export const poppins = Poppins({
   subsets: ["latin"],
