+++ pages/profile/index.js
@@ -1,18 +1,25 @@
+// Style Imports
+import { StyledCardLarge } from "@/components/StyledComponents/StyledCard";
+import { StyledWarningText } from "@/components/EditingForm";
+import {
+  StyledHeading2,
+  StyledHeading3,
+} from "@/components/StyledComponents/ParagraphsAndHeadings";
+
+// Component Imports
 import Heading from "@/components/PageHeading";
-import { MainContent, Spacer } from "@/components/StyledComponents/MainContent";
+import { Spacer } from "@/components/StyledComponents/MainContent";
+import BarDiagram from "@/components/BarDiagram";
+import UserData from "@/components/UserData";
+import SingleBarDiagram from "@/components/SingleBarDiagram";
+
+// Function and Dependency Imports
 import { hasToken } from "@/utils/checkUser";
 import {
   getAnswerOverview,
   getJLPTStats,
   getStageOverview,
 } from "@/utils/statistics";
-import BarDiagram from "@/components/BarDiagram";
-import UserData from "@/components/UserData";
-import { StyledCardLarge } from "@/components/StyledComponents/StyledCard";
-import { StyledWarningText } from "@/components/EditingForm";
-import SingleBarDiagram from "@/components/SingleBarDiagram";
-import { StyledHeading2, StyledHeading3 } from "../words/[id]";
-import PieChart from "@/components/PieChart";
 
 export async function getServerSideProps(context) {
   const token = await hasToken(context.req);
@@ -37,7 +44,7 @@ export default function ProfilePage({ wordList }) {
   const AnswerOverview = getAnswerOverview(wordList);
 
   return (
-    <MainContent>
+    <>
       <Heading>Profile</Heading>
       <StyledCardLarge>
         <UserData />
@@ -61,6 +68,6 @@ export default function ProfilePage({ wordList }) {
         />
         <Spacer />
       </StyledCardLarge>
-    </MainContent>
+    </>
   );
 }
